Build and Combined ALL files
I have created a quick java file that allows the user to combine .csv files.  With that tool I combined all the csv files and then exported it to an access database table.
The next step is to move the file to the server and export to the SQL database.