Update Server.js
Changed the localhost to look at 192.168.1.73
This is my main computer.  I was having difficulties connecting to the DB on the webserver, however I was able to have this program work on my main computer.  As a stopgap, maybe a permanent solution, I simply called the DB on my main computer.  It works and I will try to figure out why it did not on the webserver's DB.